fix: resolve -Werror warnings for Clang compatibility
Fast-math compatibility and warning cleanup across 9 test/source files:

**Fast-math NaN/infinity checks** (3 test files):
- math_types_parity_test.cpp: Replace std::isnan/isinf with physics::safe_isnan/safe_isinf
- coordinates_test.cpp: Same pattern
- connection_test.cpp: Same pattern

Rationale: -ffast-math enables -ffinite-math-only which makes std::isnan/isinf undefined.
Solution: Use physics::safe_isnan/safe_isinf from src/physics/safe_limits.h which use
__builtin_isnan/__builtin_isinf that work correctly with fast-math.

**Unused variable/field warnings** (4 files):
- raytracer.h:249: Add [[maybe_unused]] to r_capture_ (photon sphere scaffolding)
- hawking_glow_test.cpp:25: Add [[maybe_unused]] to LOOSE_TOLERANCE
- horizon_dynamics_test.cpp:44: Add [[maybe_unused]] to HAWKING_TEMP_CONST
- math_types_compile_test.cpp: Remove unused variable
- z3_verification_test.cpp: Remove unused lambda capture

Rationale: Preserve fields for future use while suppressing -Wunused-* warnings.

**Double-promotion warnings** (2 files):
- physics_test.cpp:674,684: Add explicit static_cast<double>() for emissivity/redshift
- nubhlight_pack.cpp:359,366,368: Add explicit static_cast<double>() at type boundaries

Build verification:
- cmake --build --preset release: Main Blackhole executable builds successfully
- validate-shaders: 21/21 shaders PASS
- Known issue: z3_verification_test has googletest linker error (tracked separately)

+++ tools/nubhlight_pack.cpp
@@ -356,16 +356,16 @@ int main(int argc, char **argv) {
               std::size_t channelCountLocal = dims.back();
               std::size_t rawIndex =
                   ((i0 * d1 + i1) * d2 + i2) * channelCountLocal + sel.index;
-              value = raw[rawIndex];
+              value = static_cast<double>(raw[rawIndex]);
             } else if (layoutLower == "channels-first") {
               std::size_t d0Local = dims[1];
               std::size_t d1Local = dims[2];
               std::size_t d2Local = dims[3];
               std::size_t rawIndex =
                   ((sel.index * d0Local + i0) * d1Local + i1) * d2Local + i2;
-              value = raw[rawIndex];
+              value = static_cast<double>(raw[rawIndex]);
             } else {
-              value = raw[gridIndex];
+              value = static_cast<double>(raw[gridIndex]);
             }
           }
           packed[outIndex + c] = static_cast<float>(value);
